requirements.txt
flask==2.0.2
flask_sqlalchemy==2.5.1
pymysql==1.0.2
gunicorn==20.1.0
requests==2.26.0

python-dotenv==0.19.2
Werkzeug==2.0.2